fix: refactor test helper
